pyproject.toml
[project]
name = "lang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.10",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.0",
    "langchain-community>=0.3.29",
    "langchain-huggingface>=0.1.0",
    "chromadb>=1.1.0",
    "sentence-transformers>=5.1.1",
    "pypdf>=6.1.0",
    "langchain-chroma>=0.2.6",
    "langgraph>=0.0.60",
]
